interview/app/auth/utils.py
from datetime import datetime, timedelta
import jwt
import os

SECRET_KEY = os.getenv("SECRET_KEY", "your-secret-key")


def create_access_token(data, expires_delta=timedelta(minutes=30)):
    payload = data.copy()
    payload.update({"exp": datetime.utcnow() + expires_delta})
    return jwt.encode(payload, SECRET_KEY, algorithm="HS256")
