interview/requirements.txt
Flask~=3.1.0
Flask-SQLAlchemy
Werkzeug~=3.1.3
PyJWT~=2.10.1
marshmallow~=3.23.2
python-dotenv
itsdangerous~=2.2.0